+++ lannerpsp/__init__.py
@@ -26,13 +26,16 @@ from .sdk_gps import GPS
 from .sdk_gsr import GSR, GSRDataModel, GSROffsetModel
 from .sdk_hwm import HWM, HWMSensorModel
 from .sdk_lcm import LCM
-from .sdk_odm import COMPort, COMPortInfoModel
+from .sdk_odm_com_port import COMPort, COMPortInfoModel
 from .sdk_rfm import RFM
-from .sdk_sled import GPSStatusLED, LTEStatusLED, LTEStressLED, SystemLED
+from .sdk_sled import SystemLED
+from .sdk_sled_gps import GPSStatusLED
+from .sdk_sled_lte import LTEStatusLED
+from .sdk_sled_lte_stress import LTEStressLED
 from .sdk_swr import SWR
 from .sdk_wdt import WDT, WDTInfoModel
 
-__version__ = "0.0.8"
+__version__ = "0.0.9"
 __all__ = [
     # Functions
     "get_psp_exc_msg",