+++ lannerpsp/sdk_gsr.py
@@ -17,6 +17,9 @@ from .sdk_dll import DLL
 
 logger = logging.getLogger(__name__)
 
+SUPPORTED_PLATFORMS = ("V3S", "V6S",)
+UNSUPPORTED_PLATFORMS = ("LEB-7242", "LEC-7230", "NCA-2510",)
+
 
 class GSRDataModel(NamedTuple):
     """To store G-Sensor data."""
@@ -47,10 +50,22 @@ class GSROffsetModel(NamedTuple):
 class GSR:
     """
     G-Sensor.
+
+    :param bool check_platform:
+        Set to :data:`True` to check if the platform supports this feature.
+        Defaults to :data:`False` for better compatibility.
     """
 
-    def __init__(self) -> None:
+    def __init__(self, check_platform: bool = False) -> None:
         self._version = DLL().get_version()
+        if not check_platform:
+            return
+        if self._version.platform_id in SUPPORTED_PLATFORMS:
+            pass
+        elif self._version.platform_id in UNSUPPORTED_PLATFORMS:
+            raise PSPNotSupport("Not supported on this platform")
+        else:
+            raise NotImplementedError
 
     def get_data(self) -> GSRDataModel:
         """