+++ lannerpsp/sdk_lcm.py
@@ -22,6 +22,9 @@ from .sdk_dll import DLL
 
 logger = logging.getLogger(__name__)
 
+SUPPORTED_PLATFORMS = ("NCA-2510",)
+UNSUPPORTED_PLATFORMS = ("LEB-7242", "LEC-7230", "V3S", "V6S",)
+
 DEFAULT_LCM_PORT = "/dev/ttyS1"
 DEFAULT_BAUD_RATE = 19200
 
@@ -32,12 +35,24 @@ LCM_LPT_TYPE = 2
 class LCM:
     """
     LCD Module.
+
+    :param bool check_platform:
+        Set to :data:`True` to check if the platform supports this feature.
+        Defaults to :data:`False` for better compatibility.
     """
 
-    def __init__(self) -> None:
-        self._version = DLL().get_version()
+    def __init__(self, check_platform: bool = False) -> None:
         self._str_lcm_port = c_char_p(DEFAULT_LCM_PORT.encode())
         self._dw_speed = DEFAULT_BAUD_RATE
+        self._version = DLL().get_version()
+        if not check_platform:
+            return
+        if self._version.platform_id in SUPPORTED_PLATFORMS:
+            pass
+        elif self._version.platform_id in UNSUPPORTED_PLATFORMS:
+            raise PSPNotSupport("Not supported on this platform")
+        else:
+            raise NotImplementedError
 
     def search_port(self) -> str:
         """