+++ lannerpsp/sdk_sled_gps.py
@@ -1,5 +1,4 @@
 import logging
-from ctypes import byref, c_uint8
 
 from .core import PSP, get_psp_exc_msg
 from .exc import (
@@ -19,15 +18,13 @@ from .utils import show_delay
 
 logger = logging.getLogger(__name__)
 
-SUPPORTED_PLATFORMS = ("LEB-7242", "NCA-2510",)
-UNSUPPORTED_PLATFORMS = ("LEC-7230", "V3S", "V6S",)
+SUPPORTED_PLATFORMS = ("LEB-7242",)
+UNSUPPORTED_PLATFORMS = ("LEC-7230", "NCA-2510", "V3S", "V6S",)
 
 
-class SystemLED:
+class GPSStatusLED:
     """
-    System LED.
-
-    The LED colors is dependent on the platform hardware configuration.
+    GPS Status LED.
 
     :param bool check_platform:
         Set to :data:`True` to check if the platform supports this feature.
@@ -45,46 +42,16 @@ class SystemLED:
         else:
             raise NotImplementedError
 
-    def get_status(self) -> int:
-        """
-        Get the system LED display mode status.
-
-        Example:
-
-        .. code-block:: pycon
-
-            >>> system_led = SystemLED()
-            >>> system_led.get_status()
-
-        :return: system LED status
-        :rtype: int
-        :raises PSPNotOpened: The library is not ready or opened yet.
-        :raises PSPNotSupport: This function is not supported.
-        :raises PSPError: General PSP functional error.
-        """
-        ub_read = c_uint8(0xFF)
-        with PSP() as psp:
-            i_ret = psp.lib.LMB_SLED_GetSystemLED(byref(ub_read))
-        msg = get_psp_exc_msg("LMB_SLED_GetSystemLED", i_ret)
-        if i_ret == ERR_Success:
-            return ub_read.value
-        elif i_ret == ERR_NotOpened:
-            raise PSPNotOpened(msg)
-        elif i_ret == ERR_NotSupport:
-            raise PSPNotSupport(msg)
-        else:
-            raise PSPError(msg)
-
     def off(self) -> None:
         """
-        Set the system LED display mode to off.
+        Set the GPS status LED display mode to off.
 
         Example:
 
         .. code-block:: pycon
 
-            >>> system_led = SystemLED()
-            >>> system_led.off()
+            >>> gps_status_led = GPSStatusLED()
+            >>> gps_status_led.off()
 
         :raises PSPNotOpened: The library is not ready or opened yet.
         :raises PSPInvalid: The input parameter is out of range.
@@ -92,10 +59,10 @@ class SystemLED:
         :raises PSPError: General PSP functional error.
         """
         with PSP() as psp:
-            i_ret = psp.lib.LMB_SLED_SetSystemLED(0)
-        msg = get_psp_exc_msg("LMB_SLED_SetSystemLED", i_ret)
+            i_ret = psp.lib.LMB_SLED_SetGPSLED(0)
+        msg = get_psp_exc_msg("LMB_SLED_SetGPSLED", i_ret)
         if i_ret == ERR_Success:
-            pass
+            logger.debug("set gps led off")
         elif i_ret == ERR_NotOpened:
             raise PSPNotOpened(msg)
         elif i_ret == ERR_Invalid:
@@ -104,22 +71,17 @@ class SystemLED:
             raise PSPNotSupport(msg)
         else:
             raise PSPError(msg)
-        # Check setting.
-        if self.get_status() == 0:
-            logger.debug("set status led off")
-        else:
-            logger.warning("set status led failure")
 
-    def green(self) -> None:
+    def on(self) -> None:
         """
-        Set the system LED display mode to green.
+        Set the GPS status LED display mode to on.
 
         Example:
 
         .. code-block:: pycon
 
-            >>> system_led = SystemLED()
-            >>> system_led.green()
+            >>> gps_status_led = GPSStatusLED()
+            >>> gps_status_led.on()
 
         :raises PSPNotOpened: The library is not ready or opened yet.
         :raises PSPInvalid: The input parameter is out of range.
@@ -127,10 +89,10 @@ class SystemLED:
         :raises PSPError: General PSP functional error.
         """
         with PSP() as psp:
-            i_ret = psp.lib.LMB_SLED_SetSystemLED(1)
-        msg = get_psp_exc_msg("LMB_SLED_SetSystemLED", i_ret)
+            i_ret = psp.lib.LMB_SLED_SetGPSLED(1)
+        msg = get_psp_exc_msg("LMB_SLED_SetGPSLED", i_ret)
         if i_ret == ERR_Success:
-            pass
+            logger.debug("set gps led on")
         elif i_ret == ERR_NotOpened:
             raise PSPNotOpened(msg)
         elif i_ret == ERR_Invalid:
@@ -139,22 +101,17 @@ class SystemLED:
             raise PSPNotSupport(msg)
         else:
             raise PSPError(msg)
-        # Check setting.
-        if self.get_status() == 1:
-            logger.debug("set status led green")
-        else:
-            logger.warning("set status led failure")
 
-    def red(self) -> None:
+    def blink(self) -> None:
         """
-        Set the system LED display mode to red/amber.
+        Set the GPS status LED display mode to blink.
 
         Example:
 
         .. code-block:: pycon
 
-            >>> system_led = SystemLED()
-            >>> system_led.red()
+            >>> gps_status_led = GPSStatusLED()
+            >>> gps_status_led.blink()
 
         :raises PSPNotOpened: The library is not ready or opened yet.
         :raises PSPInvalid: The input parameter is out of range.
@@ -162,10 +119,10 @@ class SystemLED:
         :raises PSPError: General PSP functional error.
         """
         with PSP() as psp:
-            i_ret = psp.lib.LMB_SLED_SetSystemLED(2)
-        msg = get_psp_exc_msg("LMB_SLED_SetSystemLED", i_ret)
+            i_ret = psp.lib.LMB_SLED_SetGPSLED(2)
+        msg = get_psp_exc_msg("LMB_SLED_SetGPSLED", i_ret)
         if i_ret == ERR_Success:
-            pass
+            logger.debug("set gps led blink")
         elif i_ret == ERR_NotOpened:
             raise PSPNotOpened(msg)
         elif i_ret == ERR_Invalid:
@@ -174,11 +131,6 @@ class SystemLED:
             raise PSPNotSupport(msg)
         else:
             raise PSPError(msg)
-        # Check setting.
-        if self.get_status() == 2:
-            logger.debug("set status led red/amber")
-        else:
-            logger.warning("set status led failure")
 
     def test(self, secs: int = 2) -> None:
         """
@@ -188,13 +140,13 @@ class SystemLED:
 
         .. code-block:: pycon
 
-            >>> system_led = SystemLED()
-            >>> system_led.test()
-            LMB_SLED_SetSystemLED: 0xffffffff: function failure
+            >>> gps_status_led = GPSStatusLED()
+            >>> gps_status_led.test()
+            set gps led on
             2. 1. 0.
-            set status led red/amber
+            set gps led blink
             2. 1. 0.
-            set status led off
+            set gps led off
 
         :param int secs: seconds for test
         :raises TypeError: The input parameters type error.
@@ -208,27 +160,27 @@ class SystemLED:
             raise ValueError("'secs' value must be >= 0")
         # Run.
         with PSP() as psp:
-            i_ret = psp.lib.LMB_SLED_SetSystemLED(1)
+            i_ret = psp.lib.LMB_SLED_SetGPSLED(1)
             if i_ret != ERR_Success:
-                msg = get_psp_exc_msg("LMB_SLED_SetSystemLED", i_ret)
+                msg = get_psp_exc_msg("LMB_SLED_SetGPSLED", i_ret)
                 print(f"\033[1;31m{msg}\033[0m")
             else:
-                print("set status led green")
+                print("set gps led on")
 
             show_delay(secs)
 
-            i_ret = psp.lib.LMB_SLED_SetSystemLED(2)
+            i_ret = psp.lib.LMB_SLED_SetGPSLED(2)
             if i_ret != ERR_Success:
-                msg = get_psp_exc_msg("LMB_SLED_SetSystemLED", i_ret)
+                msg = get_psp_exc_msg("LMB_SLED_SetGPSLED", i_ret)
                 print(f"\033[1;31m{msg}\033[0m")
             else:
-                print("set status led red/amber")
+                print("set gps led blink")
 
             show_delay(secs)
 
-            i_ret = psp.lib.LMB_SLED_SetSystemLED(0)
+            i_ret = psp.lib.LMB_SLED_SetGPSLED(0)
             if i_ret != ERR_Success:
-                msg = get_psp_exc_msg("LMB_SLED_SetSystemLED", i_ret)
+                msg = get_psp_exc_msg("LMB_SLED_SetGPSLED", i_ret)
                 print(f"\033[1;31m{msg}\033[0m")
             else:
-                print("set status led off")
+                print("set gps led off")